src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/access-key-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
Tests to ensure that accesskey works in regard to shadow DOM boundary. Can only run within DRT.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
